app/requirements.txt
faiss-cpu
fastapi
uvicorn
python-dotenv
pydantic
gunicorn
requests
beautifulsoup4